requirement.txt
flask
Flask-SQLAlchemy
flask-login
requests
pymysql